requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
boto3==1.15.18
botocore==1.18.18
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.6
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
django-import-export==2.5.0
django-storages==1.10.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
jmespath==0.10.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.0.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.25.10
xlrd==1.2.0
xlwt==1.3.0